-

CALCEPH.compute — Function.

compute(eph,jd0,time,target,center)

Compute position and velocity of target with respect to center at epoch jd0+time. This method does not support the NAIF numbering scheme. To get the best precision for the interpolation, the time is splitted in two floating-point numbers. The argument jd0 should be an integer and time should be a fraction of the day. But you may call this function with time=0 and jd0, the desired time, if you don't take care about precision.

This method does not support the NAIF body identification scheme.

Output units are:

AU and AU/day for position and velocity

rad and rad/day for librations

second and unitless for time ephemeris and time ephemeris rate

Arguments

eph: ephemeris

jd0::Float64: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris (usually TDB or TCB)

time::Float64: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris (usually TDB or TCB)

target::Integer: The body or reference point whose coordinates are required.

center::Integer: The origin of the coordinate system.

The possible values for target and center are :

1 : Mercury Barycenter

2 : Venus Barycenter

3 : Earth

4 : Mars Barycenter

5 : Jupiter Barycenter

6 : Saturn Barycenter

7 : Uranus Barycenter

8 : Neptune Barycenter

9 : Pluto Barycenter

10 : Moon

11 : Sun

12 : Solar Sytem barycenter

13 : Earth-moon barycenter

14 : Nutation angles

15 : Librations

16 : TT-TDB

17 : TCG-TCB

asteroid number + 2000000 : asteroid

source

compute(eph,jd0,time,target,center,unit)

Compute position and velocity of target with respect to center at epoch jd0+time. To get the best precision for the interpolation, the time is splitted in two floating-point numbers. The argument jd0 should be an integer and time should be a fraction of the day. But you may call this function with time=0 and jd0, the desired time, if you don't take care about precision.

Arguments

eph: ephemeris

jd0::Float64: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris (usually TDB or TCB)

time::Float64: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris (usually TDB or TCB)

target::Integer: The body or reference point whose coordinates are required. The numbering system depends on the parameter unit.

center::Integer: The origin of the coordinate system. The numbering system depends on the parameter unit.

unit::Integer : The units of the result. This integer is a sum of some unit constants (unit*) and/or the constant useNaifId. If the unit contains useNaifId, the NAIF identification numbering system is used for the target and the center. If the unit does not contain useNaifId, the old number system is used for the target and the center.

source

compute(eph,jd0,time,target,center,unit,order)

Compute position and derivatives up to order of target with respect to center at epoch jd0+time. To get the best precision for the interpolation, the time is splitted in two floating-point numbers. The argument jd0 should be an integer and time should be a fraction of the day. But you may call this function with time=0 and jd0, the desired time, if you don't take care about precision.

Arguments

eph: ephemeris

jd0::Float64: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris (usually TDB or TCB)

time::Float64: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris (usually TDB or TCB)

target::Integer: The body or reference point whose coordinates are required. The numbering system depends on the parameter unit.

center::Integer: The origin of the coordinate system. The numbering system depends on the parameter unit.

unit::Integer : The units of the result. This integer is a sum of some unit constants (unit*) and/or the constant useNaifId. If the unit contains useNaifId, the NAIF identification numbering system is used for the target and the center. If the unit does not contain useNaifId, the old number system is used for the target and the center.

order::Integer : The order of derivatives

0: only the position is computed.

1: only the position and velocity are computed.

2: only the position, velocity and acceleration are computed.

3: the position, velocity and acceleration and jerk are computed.

source

CALCEPH.unsafe_compute! — Function.

unsafe_compute!(eph,jd0,time,target,center)

In place version of the compute function. Does not perform any checks! Compute position and velocity of target with respect to center at epoch jd0+time. This method does not support the NAIF numbering scheme. To get the best precision for the interpolation, the time is splitted in two floating-point numbers. The argument jd0 should be an integer and time should be a fraction of the day. But you may call this function with time=0 and jd0, the desired time, if you don't take care about precision.

This method does not support the NAIF body identification scheme.

Output units are:

AU and AU/day for position and velocity

rad and rad/day for librations

second and unitless for time ephemeris and time ephemeris rate

Arguments

result: container for result. It is not checked if it is sufficiently large enough!

eph: ephemeris

jd0::Float64: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris (usually TDB or TCB)

time::Float64: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris (usually TDB or TCB)

target::Integer: The body or reference point whose coordinates are required.

center::Integer: The origin of the coordinate system.

Return:

status integer from CALCEPH: 0 if an error occured

The possible values for target and center are :

1 : Mercury Barycenter

2 : Venus Barycenter

3 : Earth

4 : Mars Barycenter

5 : Jupiter Barycenter

6 : Saturn Barycenter

7 : Uranus Barycenter

8 : Neptune Barycenter

9 : Pluto Barycenter

10 : Moon

11 : Sun

12 : Solar Sytem barycenter

13 : Earth-moon barycenter

14 : Nutation angles

15 : Librations

16 : TT-TDB

17 : TCG-TCB

asteroid number + 2000000 : asteroid

source

unsafe_compute!(result,eph,jd0,time,target,center,unit)

In place version of the compute function. Does not perform any checks! Compute position and velocity of target with respect to center at epoch jd0+time. To get the best precision for the interpolation, the time is splitted in two floating-point numbers. The argument jd0 should be an integer and time should be a fraction of the day. But you may call this function with time=0 and jd0, the desired time, if you don't take care about precision.

Arguments

result: container for result. It is not checked if it is sufficiently large enough!

eph: ephemeris

jd0::Float64: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris (usually TDB or TCB)

time::Float64: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris (usually TDB or TCB)

target::Integer: The body or reference point whose coordinates are required. The numbering system depends on the parameter unit.

center::Integer: The origin of the coordinate system. The numbering system depends on the parameter unit.

unit::Integer : The units of the result. This integer is a sum of some unit constants (unit*) and/or the constant useNaifId. If the unit contains useNaifId, the NAIF identification numbering system is used for the target and the center. If the unit does not contain useNaifId, the old number system is used for the target and the center.

Return:

status integer from CALCEPH: 0 if an error occured

source

unsafe_compute!(result,eph,jd0,time,target,center,unit,order)

In place version of the compute function. Does not perform any checks! Compute position and derivatives up to order of target with respect to center at epoch jd0+time. To get the best precision for the interpolation, the time is splitted in two floating-point numbers. The argument jd0 should be an integer and time should be a fraction of the day. But you may call this function with time=0 and jd0, the desired time, if you don't take care about precision.

Arguments

result: container for result. It is not checked if it is sufficiently large enough!

eph: ephemeris

jd0::Float64: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris (usually TDB or TCB)

time::Float64: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris (usually TDB or TCB)

target::Integer: The body or reference point whose coordinates are required. The numbering system depends on the parameter unit.

center::Integer: The origin of the coordinate system. The numbering system depends on the parameter unit.

unit::Integer : The units of the result. This integer is a sum of some unit constants (unit*) and/or the constant useNaifId. If the unit contains useNaifId, the NAIF identification numbering system is used for the target and the center. If the unit does not contain useNaifId, the old number system is used for the target and the center.

order::Integer : The order of derivatives

0: only the position is computed.

1: only the position and velocity are computed.

2: only the position, velocity and acceleration are computed.

3: the position, velocity and acceleration and jerk are computed.

Return:

status integer from CALCEPH: 0 if an error occured

source

CALCEPH.orient — Function.

orient(eph,jd0,time,target,unit)

Compute Euler angles and first derivative for the orientation of target at epoch jd0+time. To get the best precision for the interpolation, the time is split in two floating-point numbers. The argument jd0 should be an integer and time should be a fraction of the day. But you may call this function with time=0 and jd0, the desired time, if you don't take care about precision.

Arguments

eph: ephemeris

jd0::Float64: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris (usually TDB or TCB)

time::Float64: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris (usually TDB or TCB)

target::Integer: The body whose orientation is required. The numbering system depends on the parameter unit.

unit::Integer : The units of the result. This integer is a sum of some unit constants (unit*) and/or the constant useNaifId. If the unit contains useNaifId, the NAIF identification numbering system is used for the target and the center. If the unit does not contain useNaifId, the old number system is used for the target and the center (see the list in the documentation of function compute). The angles are expressed in radians if unit contains unitRad. If the unit contains outputNutationAngles, the nutation angles are computed rather than the Euler angles.

source

orient(eph,jd0,time,target,unit,order)

Compute Euler angles and derivatives up to order for the orientation of target at epoch jd0+time. To get the best precision for the interpolation, the time is split in two floating-point numbers. The argument jd0 should be an integer and time should be a fraction of the day. But you may call this function with time=0 and jd0, the desired time, if you don't take care about precision.

Arguments

eph: ephemeris

jd0::Float64: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris (usually TDB or TCB)

time::Float64: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris (usually TDB or TCB)

target::Integer: The body whose orientation is required. The numbering system depends on the parameter unit.

unit::Integer : The units of the result. This integer is a sum of some unit constants (unit*) and/or the constant useNaifId. If the unit contains useNaifId, the NAIF identification numbering system is used for the target and the center. If the unit does not contain useNaifId, the old number system is used for the target and the center (see the list in the documentation of function compute). If the unit contains outputNutationAngles, the nutation angles are computed rather than the Euler angles.

order::Integer : The order of derivatives

0: only the angles are computed.

1: only the angles and 1st derivatives are computed.

2: only the angles, the 1st derivatives and 2nd derivatives are computed.

3: the angles, the 1st derivatives, 2nd derivatives and 3rd derivatives are computed.

source

CALCEPH.unsafe_orient! — Function.

unsafe_orient!(result,eph,jd0,time,target,unit)

In place version of the orient function. Does not perform any checks! Compute Euler angles and first derivative for the orientation of target at epoch jd0+time. To get the best precision for the interpolation, the time is split in two floating-point numbers. The argument jd0 should be an integer and time should be a fraction of the day. But you may call this function with time=0 and jd0, the desired time, if you don't take care about precision.

Arguments

result: container for result. It is not checked if it is sufficiently large enough!

eph: ephemeris

jd0::Float64: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris (usually TDB or TCB)

time::Float64: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris (usually TDB or TCB)

target::Integer: The body whose orientation is required. The numbering system depends on the parameter unit.

unit::Integer : The units of the result. This integer is a sum of some unit constants (unit*) and/or the constant useNaifId. If the unit contains useNaifId, the NAIF identification numbering system is used for the target and the center. If the unit does not contain useNaifId, the old number system is used for the target and the center (see the list in the documentation of function compute). If the unit contains outputNutationAngles, the nutation angles are computed rather than the Euler angles.

Return:

status integer from CALCEPH: 0 if an error occured

source

unsafe_orient!(result,eph,jd0,time,target,unit,order)

In place version of the orient function. Does not perform any checks! Compute Euler angles and derivatives up to order for the orientation of target at epoch jd0+time. To get the best precision for the interpolation, the time is split in two floating-point numbers. The argument jd0 should be an integer and time should be a fraction of the day. But you may call this function with time=0 and jd0, the desired time, if you don't take care about precision.

Arguments

result: container for result. It is not checked if it is sufficiently large enough!

eph: ephemeris

jd0::Float64: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris (usually TDB or TCB)

time::Float64: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris (usually TDB or TCB)

target::Integer: The body whose orientation is required. The numbering system depends on the parameter unit.

unit::Integer : The units of the result. This integer is a sum of some unit constants (unit*) and/or the constant useNaifId. If the unit contains useNaifId, the NAIF identification numbering system is used for the target and the center. If the unit does not contain useNaifId, the old number system is used for the target and the center (see the list in the documentation of function compute). If the unit contains outputNutationAngles, the nutation angles are computed rather than the Euler angles.

order::Integer : The order of derivatives

0: only the angles are computed.

1: only the angles and 1st derivatives are computed.

2: only the angles, the 1st derivatives and 2nd derivatives are computed.

3: the angles, the 1st derivatives, 2nd derivatives and 3rd derivatives are computed.

Return:

status integer from CALCEPH: 0 if an error occured

source

CALCEPH.rotAngMom — Function.

rotAngMom(eph,jd0,time,target,unit)

Compute angular momentum due to rotation and first derivative of target at epoch jd0+time. To get the best precision for the interpolation, the time is splitted in two floating-point numbers. The argument jd0 should be an integer and time should be a fraction of the day. But you may call this function with time=0 and jd0, the desired time, if you don't take care about precision.

Arguments

eph: ephemeris

jd0::Float64: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris (usually TDB or TCB)

time::Float64: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris (usually TDB or TCB)

target::Integer: The body whose angular momentum is required. The numbering system depends on the parameter unit.

unit::Integer : The units of the result. This integer is a sum of some unit constants (unit*) and/or the constant useNaifId. If the unit contains useNaifId, the NAIF identification numbering system is used for the target and the center. If the unit does not contain useNaifId, the old number system is used for the target and the center (see the list in the documentation of function compute). The angles are expressed in radians if unit contains unitRad.

source

rotAngMom(eph,jd0,time,target,unit,order)

Compute angular momentum due to rotation and derivatives up to order of target at epoch jd0+time. To get the best precision for the interpolation, the time is splitted in two floating-point numbers. The argument jd0 should be an integer and time should be a fraction of the day. But you may call this function with time=0 and jd0, the desired time, if you don't take care about precision.

Arguments

eph: ephemeris

jd0::Float64: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris (usually TDB or TCB)

time::Float64: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris (usually TDB or TCB)

target::Integer: The body whose angular momentum is required. The numbering system depends on the parameter unit.

unit::Integer : The units of the result. This integer is a sum of some unit constants (unit*) and/or the constant useNaifId. If the unit contains useNaifId, the NAIF identification numbering system is used for the target and the center. If the unit does not contain useNaifId, the old number system is used for the target and the center (see the list in the documentation of function compute).

order::Integer : The order of derivatives

0: only the angles are computed.

1: only the angles and 1st derivatives are computed.

2: only the angles, the 1st derivatives and 2nd derivatives are computed.

3: the angles, the 1st derivatives, 2nd derivatives and 3rd derivatives are computed.

source

CALCEPH.unsafe_rotAngMom! — Function.

unsafe_rotAngMom!(result,eph,jd0,time,target,unit)

In place version of the rotAngMom function. Does not perform any checks! Compute angular momentum due to rotation and first derivative of target at epoch jd0+time. To get the best precision for the interpolation, the time is splitted in two floating-point numbers. The argument jd0 should be an integer and time should be a fraction of the day. But you may call this function with time=0 and jd0, the desired time, if you don't take care about precision.

Arguments

result: container for result. It is not checked if it is sufficiently large enough!

eph: ephemeris

jd0::Float64: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris (usually TDB or TCB)

time::Float64: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris (usually TDB or TCB)

target::Integer: The body whose angular momentum is required. The numbering system depends on the parameter unit.

unit::Integer : The units of the result. This integer is a sum of some unit constants (unit*) and/or the constant useNaifId. If the unit contains useNaifId, the NAIF identification numbering system is used for the target and the center. If the unit does not contain useNaifId, the old number system is used for the target and the center (see the list in the documentation of function compute). The angles are expressed in radians if unit contains unitRad.

Return:

status integer from CALCEPH: 0 if an error occured

source

unsafe_rotAngMom!(result,eph,jd0,time,target,unit,order)

In place version of the rotAngMom function. Does not perform any checks! Compute angular momentum due to rotation and derivatives up to order of target at epoch jd0+time. To get the best precision for the interpolation, the time is splitted in two floating-point numbers. The argument jd0 should be an integer and time should be a fraction of the day. But you may call this function with time=0 and jd0, the desired time, if you don't take care about precision.

Arguments

result: container for result. It is not checked if it is sufficiently large enough!

eph: ephemeris

jd0::Float64: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris (usually TDB or TCB)

time::Float64: jd0+time must be equal to the Julian Day for the time coordinate corresponding to the ephemeris (usually TDB or TCB)

target::Integer: The body whose angular momentum is required. The numbering system depends on the parameter unit.

unit::Integer : The units of the result. This integer is a sum of some unit constants (unit*) and/or the constant useNaifId. If the unit contains useNaifId, the NAIF identification numbering system is used for the target and the center. If the unit does not contain useNaifId, the old number system is used for the target and the center (see the list in the documentation of function compute).

order::Integer : The order of derivatives

0: only the angles are computed.

1: only the angles and 1st derivatives are computed.

2: only the angles, the 1st derivatives and 2nd derivatives are computed.

3: the angles, the 1st derivatives, 2nd derivatives and 3rd derivatives are computed.

Return:

status integer from CALCEPH: 0 if an error occured

source