-

CALCEPH.BodyId — Type.

BodyId

Body identifiers.

source

CALCEPH.add! — Function.

add!(bid,name,id)

Add a new mapping name->id into BodyId instance bid.

Example:

    using CALCEPH
    bid=CALCEPH.BodyId()
    CALCEPH.add!(bid,:tatooine,1000001)
    CALCEPH.add!(bid,:dagobah,1000002)
    CALCEPH.add!(bid,:endor,1000003)
    CALCEPH.add!(bid,:deathstar,1000004)
    CALCEPH.add!(bid,:endor_deathstar_system_barycenter,1000005)
    CALCEPH.add!(bid,:edsb,1000005)

# output

source

CALCEPH.loadData! — Function.

loadData!(bid,filename)

Load mapping (body name,body ID) from file into BodyId instance bid. Names from the file are converted to lower case and have spaces replaced by underscores before being converted to symbols/interned strings.

Example file https://github.com/bgodard/CALCEPH.jl/blob/master/data/NaifIds.txt

source

CALCEPH.naifId — Constant.

naifId

NAIF identification numbers

Examples:

julia> using CALCEPH

julia> naifId.id[:sun]
10

julia> naifId.id[:mars]
499

julia> naifId.names[0]
Set(Symbol[:ssb, :solar_system_barycenter])

source