-

CALCEPH.unitAU — Constant.

unitAU

Astronomical unit: distance unit

source

CALCEPH.unitKM — Constant.

unitKM

kilometer: distance unit

source

CALCEPH.unitDay — Constant.

unitDay

day: time unit

source

CALCEPH.unitSec — Constant.

unitSec

second: time unit

source

CALCEPH.unitRad — Constant.

unitRad

radian: angle unit

source

CALCEPH.useNaifId — Constant.

useNaifId

has to be added to the unit argument when using NAIF integer codes for identification of center and target

source

CALCEPH.outputEulerAngles — Constant.

outputEulerAngles

has to be added to the unit argument for orient to output Euler angles

source

CALCEPH.outputNutationAngles — Constant.

outputNutationAngles

has to be added to the unit argument for orient to output nutation angles

source